Use different API for loading datasets and create branching logic for creating indexes
